Refactoring

+++ fivehundredpx-example/src/main/java/com/eharmony/example/service/FiveHundredPxSpiceRequest.java
@@ -9,10 +9,10 @@ import com.eharmony.example.model.FiveHundredPxPhotoContainer;
  */
 public class FiveHundredPxSpiceRequest extends RetrofitSpiceRequest<FiveHundredPxPhotoContainer, FiveHundredPx>{
 
-    private int page;
+    private long page;
     private int resultPerPage;
 
-    public FiveHundredPxSpiceRequest(final int page, final int resultPerPage) {
+    public FiveHundredPxSpiceRequest(final long page, final int resultPerPage) {
         super(FiveHundredPxPhotoContainer.class, FiveHundredPx.class);
         this.page = page;
         this.resultPerPage = resultPerPage;
